make restoring from cni the default base case and remove the kube variant
