+++ cns/podprovider/cni/podinfoprovider_test.go
@@ -1,12 +1,9 @@
-package podprovider
+package cni
 
 import (
-	"net"
 	"testing"
 
 	"github.com/Azure/azure-container-networking/cns"
-	"github.com/Azure/azure-container-networking/cns/restserver"
-	"github.com/Azure/azure-container-networking/store"
 	testutils "github.com/Azure/azure-container-networking/test/utils"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/utils/exec"
@@ -52,53 +49,6 @@ func TestNewCNIPodInfoProvider(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			got, err := newCNIPodInfoProvider(tt.exec)
-			if tt.wantErr {
-				assert.Error(t, err)
-				return
-			}
-			assert.NoError(t, err)
-			podInfoByIP, _ := got.PodInfoByIP()
-			assert.Equal(t, tt.want, podInfoByIP)
-		})
-	}
-}
-
-func TestNewCNSPodInfoProvider(t *testing.T) {
-	goodStore := store.NewMockStore("")
-	goodEndpointState := make(map[string]*restserver.EndpointInfo)
-	endpointInfo := &restserver.EndpointInfo{PodName: "goldpinger-deploy-bbbf9fd7c-z8v4l", PodNamespace: "default", IfnameToIPMap: make(map[string]*restserver.IPInfo)}
-	endpointInfo.IfnameToIPMap["eth0"] = &restserver.IPInfo{IPv4: []net.IPNet{{IP: net.IPv4(10, 241, 0, 65), Mask: net.IPv4Mask(255, 255, 255, 0)}}}
-
-	goodEndpointState["0a4917617e15d24dc495e407d8eb5c88e4406e58fa209e4eb75a2c2fb7045eea"] = endpointInfo
-	err := goodStore.Write(restserver.EndpointStoreKey, goodEndpointState)
-	if err != nil {
-		t.Fatalf("Error writing to store: %v", err)
-	}
-	tests := []struct {
-		name    string
-		store   store.KeyValueStore
-		want    map[string]cns.PodInfo
-		wantErr bool
-	}{
-		{
-			name:  "good",
-			store: goodStore,
-			want: map[string]cns.PodInfo{"10.241.0.65": cns.NewPodInfo("0a4917617e15d24dc495e407d8eb5c88e4406e58fa209e4eb75a2c2fb7045eea",
-				"0a4917617e15d24dc495e407d8eb5c88e4406e58fa209e4eb75a2c2fb7045eea", "goldpinger-deploy-bbbf9fd7c-z8v4l", "default")},
-			wantErr: false,
-		},
-		{
-			name:    "empty store",
-			store:   store.NewMockStore(""),
-			want:    map[string]cns.PodInfo{},
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := newCNSPodInfoProvider(tt.store)
 			if tt.wantErr {
 				assert.Error(t, err)
 				return
