+++ cns/podprovider/cni/statefile.go
@@ -1,4 +1,4 @@
-package podprovider
+package cni
 
 import (
 	"encoding/json"
